Reorganizei funcoes em diferentes arquivos e readaptei as funcoes de solucionado e nJogadas
