+++ jogada.c
@@ -1,0 +1,7 @@
+// Estrutura para jogada
+typedef struct {
+    int daLin;
+    int daCol;
+    int ateLin;
+    int ateCol;
+} Jogada;
