+++ jogada.h
@@ -1,0 +1,11 @@
+#ifndef JOGADA_H
+#define JOGADA_H
+
+typedef struct {
+    int daLin;
+    int daCol;
+    int ateLin;
+    int ateCol;
+} Jogada;
+
+#endif // JOGADA_H
