+++ matrizes_para_teste/m1.txt
@@ -1,0 +1,7 @@
+-1, -1, 1, 1, 1, -1, -1
+-1, -1, 1, 1, 1, -1, -1
+1, 1, 1, 1, 1, 1, 1
+1, 1, 1, 0, 1, 1, 1
+1, 1, 1, 1, 1, 1, 1
+-1, -1, 1, 1, 1, -1, -1
+-1, -1, 1, 1, 1, -1, -1
