+++ matrizes_para_teste/m2.txt
@@ -1,0 +1,7 @@
+-1, -1, 0, 0, 0, -1, -1
+-1, -1, 0, 0, 0, -1, -1
+0, 0, 0, 0, 0, 0, 0
+0, 0, 0, 1, 0, 0, 0
+0, 0, 0, 0, 0, 0, 0
+-1, -1, 0, 0, 0, -1, -1
+-1, -1, 0, 0, 0, -1, -1
