+++ matrizes_para_teste/m3.txt
@@ -1,0 +1,10 @@
+-1, -1, -1, -1, 1, -1, -1, -1, -1
+-1, -1, -1, 1, 1, 1, -1, -1, -1
+-1, -1, 1, 1, 1, 1, 1, -1, -1
+-1, 1, 1, 1, 1, 1, 1, 1, -1
+1, 1, 1, 1, 0, 1, 1, 1, 1
+-1, 1, 1, 1, 1, 1, 1, 1, -1
+-1, -1, 1, 1, 1, 1, 1, -1, -1
+-1, -1, -1, 1, 1, 1, -1, -1, -1
+-1, -1, -1, -1, 1, -1, -1, -1, -1
+
