+++ uteis.h
@@ -1,0 +1,10 @@
+#ifndef UTEIS_H
+#define UTEIS_H
+
+
+void printMatriz(int** matriz, int tamanho);
+void printMatrizASCII(int** matriz, int tamanho);
+int determinarTamanho(char linha[]);
+int nJogadasNecessarias(int** matriz, int tamanho);
+
+#endif
